Made major components of program modular for reading purposes (#7)
